requirements.txt
aliyun-python-sdk-core==2.13.30
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.14.0
appdirs==1.4.4
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
crcmod==1.7
cryptography==3.2.1
cssselect==1.1.0
distlib==0.3.1
filelock==3.0.12
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.4.0
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.2
oss2==2.13.1
parsel==1.6.0
Pillow==8.1.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==1.0.2
pyOpenSSL==20.0.1
queuelib==1.5.0
removebg==0.4
requests==2.25.1
Scrapy==2.4.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.1
threadpool==1.3.2
Twisted==20.3.0
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.4.0
w3lib==1.22.0
zipp==3.4.0
zope.interface==5.2.0
